# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rust Spell Server",
    "version": "0.1.0",
    "description": "A production-ready Rust HTTP server exposing a spell-checking API over OpenAPI-compliant endpoints.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "contact": {
      "name": "Brian Tafoya",
      "email": "btafoya@briantafoya.com",
      "url": "https://github.com/btafoya/rustspell-server"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "security": [],
  "paths": {
    "/api-keys": {
      "post": {
        "operationId": "createApiKey",
        "tags": ["API Keys"],
        "summary": "Create an API key",
        "description": "Creates a new `admin` or `standard` key scoped to the caller's own tenant. The raw key value is returned exactly once.",
        "security": [{ "apiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateApiKeyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Key created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatedApiKeyResponse" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, or revoked key",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "403": {
            "description": "Key does not have the admin role",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "429": {
            "description": "Too many failed authentication attempts from this IP",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listApiKeys",
        "tags": ["API Keys"],
        "summary": "List API keys",
        "description": "Returns metadata for all keys belonging to the caller's own tenant. Raw and hashed values are never returned.",
        "security": [{ "apiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Key metadata list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyListResponse" }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, or revoked key",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "403": {
            "description": "Key does not have the admin role",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "operationId": "revokeApiKey",
        "tags": ["API Keys"],
        "summary": "Revoke an API key",
        "description": "Soft-revokes a key belonging to the caller's own tenant. Idempotent: revoking an already-revoked key still returns 204.",
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Key revoked (or already revoked)" },
          "401": {
            "description": "Missing, invalid, expired, or revoked key",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "403": {
            "description": "Key does not have the admin role",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "404": {
            "description": "Unknown key id, or the id belongs to a different tenant",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}/rotate": {
      "post": {
        "operationId": "rotateApiKey",
        "tags": ["API Keys"],
        "summary": "Rotate an API key",
        "description": "Issues a new raw value for an existing key, invalidating the previous value. The id, label, and role are unchanged; the raw key is returned exactly once.",
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Key rotated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatedApiKeyResponse" }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, or revoked key",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "403": {
            "description": "Key does not have the admin role",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "404": {
            "description": "Unknown key id, or the id belongs to a different tenant",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/tenant": {
      "get": {
        "operationId": "getOwnTenant",
        "tags": ["Tenants"],
        "summary": "Get own tenant",
        "description": "Returns metadata for the calling key's own tenant. Callable by `admin` or `standard` keys (read-only self-service usage visibility).",
        "security": [{ "apiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Tenant metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TenantMetadata" }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, or revoked key",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "403": {
            "description": "Origin not registered to the caller's tenant, or the tenant is suspended",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
...
    "/tenant/origins/{id}": {
      "delete": {
        "operationId": "revokeOrigin",
        "tags": ["Origins"],
        "summary": "Revoke a registered origin",
        "description": "Removes an origin belonging to the caller's own tenant.",
        "security": [{ "apiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Origin revoked" },
          "401": {
            "description": "Missing, invalid, expired, or revoked key",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "403": {
            "description": "Key does not have the admin role, origin not registered to the caller's tenant, or the tenant is suspended",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "404": {
            "description": "Unknown origin id, or the id belongs to a different tenant",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
...
    "/docs": {
      "get": {
        "operationId": "getOpenApiSpec",
        "tags": ["Documentation"],
        "summary": "OpenAPI specification",
        "description": "Returns the OpenAPI 3.0 JSON document describing this API.",
        "responses": {
          "200": {
            "description": "OpenAPI JSON document",
            "content": {
              "application/json": {
                "schema": {
                  "description": "OpenAPI 3.0.3 root object",
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateApiKeyRequest": {
        "type": "object",
        "required": ["label", "role"],
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Human-readable name to distinguish this key from others."
          },
          "role": { "$ref": "#/components/schemas/Role" },
          "expires_at": {
            "type": "integer",
            "nullable": true,
            "description": "Unix-epoch seconds. Omit for a key that never expires."
          }
        }
      },
...
      "CreatedApiKeyResponse": {
        "description": "`ApiKeyMetadata` plus the raw key value, returned exactly once (on creation or rotation).",
        "allOf": [
          { "$ref": "#/components/schemas/ApiKeyMetadata" },
          {
            "type": "object",
            "required": ["key"],
            "properties": {
              "key": {
                "type": "string",
                "description": "Raw key value. Not persisted, not retrievable again — save it now."
              }
            }
          }
        ]
      },
      "ApiKeyListResponse": {
        "type": "object",
        "required": ["keys"],
        "properties": {
          "keys": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApiKeyMetadata" }
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}